Fix usage of volk_32f_index_max*, first argument, to be uint16_t* as required by the Volk API.

+++ src/algorithms/acquisition/gnuradio_blocks/galileo_e5a_noncoherent_iq_acquisition_caf_cc.cc
@@ -424,11 +424,11 @@ int galileo_e5a_noncoherentIQ_acquisition_caf_cc::general_work(int noutput_items
 
             // initialize acquisition algorithm
             int doppler;
-            unsigned int indext = 0;
-            unsigned int indext_IA = 0;
-            unsigned int indext_IB = 0;
-            unsigned int indext_QA = 0;
-            unsigned int indext_QB = 0;
+            uint16_t indext = 0;
+            uint16_t indext_IA = 0;
+            uint16_t indext_IB = 0;
+            uint16_t indext_QA = 0;
+            uint16_t indext_QB = 0;
             float magt = 0.0;
             float magt_IA = 0.0;
             float magt_IB = 0.0;
